Streamlined profile management
Instead of managing both a live and a temporary profile, changes are now always made to the live profile, with the Cancel button reverting to a backup copy.
Consequently, most profile management functions have lost their "profiles" parameter and now operate directly on PagingProfiles.
The use of the "Defaults" button is now supported.