Adjust position of paging indicator frame
With the changed frame reference, we also need to adjust the position of
the frame that indicates an overriden page. As far as I know, we cannot
reference MainMenuBarArtFrame.PageNumber directly with an anchor, so
instead we reference the same anchor point of its parent and adjust the
offset accordingly.

Signed-off-by: Teomyr <teo@teomyr.eu>